Validate configured keyspace name in CassandraPluginConfig
Verify that the user supplied keyspace and table name conforms to
Cassandra's keyspace name specification which states that valid
keyspace names may be composed of letters, numbers and underscores
with a max of 32 characters.

If a name doesn't comply, fail early and provide a useful error
message to the user.

Add tests for CassandraPluginConfig keyspace and table name error
validation