Don't use BATCH when a single message is persisted.
This should be the default case when the app `persist`s an event.
In our application we're persisting events with a payload > 5kb,
which triggers a warning in C* due to the default setting of
`batch_size_warn_threshold: 5k`, which will not be triggered for
`executeAsync` (if this warning is valid or not for single-partition/
single-statement batches might be clarified in the future, see also
https://issues.apache.org/jira/browse/CASSANDRA-6487).

Our ops team prefers not to increase this configuration, because in case
of C* stability issues Datastax at first points to such settings that
don't follow the recommendations.