update to Akka 2.4.1
* don't support rejections, avoid holes in seq nr
* fix bug in asyncDeleteMessagesTo (found by new TCK test)

When exception is thrown from asyncWriteMessages or if the Future is completed
with failure the AsyncWriteJournal will translate that to a WriteMessageFailed
message that will cause the PersistentActor to stop. When it is started again
it will continue from highest sequence number.

If the Try inside the Seq is a Failure the AsyncWriteJournal will translate that
to a WriteMessageRejected and the PersistentActor will happily continue with
next sequence number, i.e. there will be a hole.

Rejection is a new optional feature in Akka 2.4, which we don't want to support
in the Cassandra plugin because in eventsByTag query (and possibly other places)
we can detect missing/delayed events easily when there are no holes in the
sequence number series.