Add fork option to CassandraLauncher
Fixes #153.

Restarting an embedded Cassandra in memory is not supported, see:

https://issues.apache.org/jira/browse/CASSANDRA-11154

In there a Cassandra core developer says "Oh, and C* is definitely not
designed to be in-memory restartable". Hence, when using the
CassandraLauncher for tests, it should be forked into its own process.

This commit introduces that forking. It maintains backwards
compatibility, and keeps the default behaviour to not fork, but adds a
flag that can be set to fork.

Updated as per review comments

* Split start into smaller functions
* Made the default behaviour to fork

(cherry picked from commit 3de3b46ccd04d11513a7f61e65d471fd9af2042e)