Allow use of shaded Cassandra
This removes the binary dependency on Cassandra, and instead detects
Cassandra from the classpath, either Cassandra proper, or the shaded
version published by cassandra-unit (LGPL).

Projects can use the shaded version to avoid classpath conflicts with
Cassandra itself, such as with the version of Netty or Guice that it
depends on.