Bundled Cassandra into a fat jar on the classpath
This splits the CassandraLauncher into its own artifact. That artifact
bundles a Cassandra fat jar, which means that Cassandra and its
dependencies do not need to be on the classpath.  When Cassandra is
forked, that jar is extracted, and used as the classpath for Cassandra.

This solves the problem of Cassandra (and its many dependencies) needing
to be on the classpath, and the conflicts that can create (such as with
Netty, Guice, etc).

One issue with this is that it means you can't just add something to
your classpath to get Cassandra to use it, such as a custom log
configuration, etc. To solve this, an additional classpath parameter has
been added to the start method, which allows additional classpath
elements to be added to the classpath. A helper has also been added for
generating this parameter, that takes a list of resources (eg,
logback-test.xml), and calculates the directories/jars that they come
from, and returns those directories for adding to the classpath.