Harden events by tag stage spec
The test checks that the stage finds a first event in a previous bucket
but due to timing the test some times puts the event two buckets ago
due to the stage querying until the bucket after when the timeuuids
were generated.

This changes the test to hold off if it is within 2 seconds of a bucket
end.

Refs #278