Remove internal state for tag writer when persistent actor shuts down (#412)
* Remove internal state for tag writer when persistent actor shuts down

* Removes tag pid sequence nr state from each TagWriter
* Removes sequenceNrs which was there for debugging messages from the
journal that came out of order. This has never happend since initial
version.

* Assign tag pid sequence nrs on event arrival

Changes from calculating the tag pid sequence nr when an event is
written. Meaning that it is always safe to drop the state for a
persistence id in a TagWriter. Also means that the tag pid sequence nrs
don't need to be restored if a write fails.