Review events by tag stage logging (#458)
The existing logging is far too verbose. Logging multiple actions as
INFO when looking for a missing event.

This change reduces it to one INFO log when the problem is found, then
all the actions at DEBUG followed by one log line to indicate that
the problem is fixed or that the event was not found.