Update snapshot deletion to use range query #450
Support for range deletions in Cassandra was added in version 3.0 and
the previous operation was implemented with that limitation and is kept
to be used if the cassandra-2x-compat flag is enabled in the config.
Otherwise this new change means the default query is a more performant
single range deletion as opposed to the compound delete query used before.