Snapshots deletion optimisation #450
Support for range deletions in Cassandra was added in version 3.0 and
the previous operation was implemented within that limitation and is kept
to be used if the cassandra-2x-compat flag is enabled in the config.
Otherwise this new change means the default query is a more performant
single range deletion as opposed to the compound delete query used before.

* Snapshot row deletions grouped into blocks to avoided exceeding 65535
statement limit in a single batch query.
* Added condition to the use of the CQL range delete to only be used
when there is no timestamp constraint as the where clause cannot contain
a range constraint in the timestamp column as well as the sequence_nr
in the same query