EventsByTag current queries, use toOffset at end of current millisecond
It used to use the beginning meaning that any event written in the
current millisecond won't be picked up. This doesn't seem right
as an event that has just been written may not be picked up for a query
started after (but in the same ms).

This causes a few tests to fail infrequently.