fix: Buffer.addPending sets writeRequired (#1015)
```scala
require(pending.isEmpty, "Pending should be empty if write not required")
```

suggests that an invariant of `journal.Buffer` is that `writeRequired || pending.nonEmpty`.  The `addPending` method, if called on a `Buffer` where `writeRequired` is `false` will break this invariant, which may cause subsequent adds to the buffer to throw, crashing the `TagWriter` and forcing an `ActorSystem` restart.